[feat]: Add code for pose estimation using linear pno and extra visualization codes

+++ src/triangulation.py
@@ -20,7 +20,7 @@ def triangulate_pts(x1, x2, P1, P2):
         X_h (np.array, shape(n,4)): 3D triangulated homogeneous coordinate in space
     """
     assert len(x1) == len(x2)
-    Xh = np.zeros((0, 4))
+    Xh = np.zeros((0, 4), dtype=np.float32)
     for i in range(len(x1)):
         skew1 = Vec2Skew(x1[i])
         skew2 = Vec2Skew(x2[i])
@@ -28,5 +28,5 @@ def triangulate_pts(x1, x2, P1, P2):
         u, s, v = np.linalg.svd(A)
         X = v.T[:, -1]
         Xn = X / X[-1]
-        Xh = np.vstack((Xh, Xn))
+        Xh = np.vstack((Xh, Xn), dtype=np.float32)
     return Xh
